wip: Move Add Model from Modal to new Panel
